refactor: Input Component 수정
